Simplify test framework.
Previously the test framework had somewhat redundant concepts,
like namespaces and categories, and was pretty complicated.

Now there's just a category (corresponding to a subdirectory of
`tests`), and a test runner simply accepts a test as a `&str` and
returns a `String`. A test fails if it panics or if its output
doesn't match the previously generated output. An "expected failure"
can be handled by the test runner returning a string representing
the error(s).

Compiler tests no longer emit hashes of the AST. This made our
commits a lot bigger and was an extra pain when changing passes that
ultimately led to the same code output. In the future if it seems
like it'd be useful I'll implement another easy way to compare pass
outputs from different runs of the compiler.

Also:

* Compiler tests now use the pass structure from the compiler,
rather than duplicating it.

* The test framework is now in the base directory of the repo.

* Manually implemented (de)serialization for Identifier and
ProgramId is gone.

* Serialization tests are gone.